Backend/fastapi-backend/requirements.txt
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.29.0,<0.30.0 # Includes websockets support
python-dotenv>=1.0.1,<2.0.0
neo4j>=5.18.0,<6.0.0 # From langchain_arch requirements
pandas>=2.0.0,<3.0.0 # Likely needed by langchain_arch or for data handling

# Add LangChain related dependencies (pin versions based on langchain_arch/requirements.txt)
langchain
langchain-community
langchain-openai
langchain_core
openai>=1.14.0,<2.0.0
e2b-code-interpreter
# Add other dependencies if needed by your langchain_arch module
# e.g., nest-asyncio if it's required by underlying libs in this async context
nest-asyncio>=1.6.0,<2.0.0 
plotly>=5.0.0,<6.0.0

langchain-text-splitters
mem0ai

